+++ KoS.Apps.SharePoint.SmartCAML/KoS.Apps.SharePoint.SmartCAML.Editor/UserControls/QueryBuilderControl.cs
@@ -5,6 +5,7 @@ using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Data;
 using KoS.Apps.SharePoint.SmartCAML.Editor.Builder;
+using KoS.Apps.SharePoint.SmartCAML.Editor.Core;
 
 namespace KoS.Apps.SharePoint.SmartCAML.Editor.Controls
 {
@@ -14,6 +15,7 @@ namespace KoS.Apps.SharePoint.SmartCAML.Editor.Controls
     public partial class QueryBuilderControl : UserControl
     {
         public event EventHandler Changed;
+        public OrderedList<QueryFilterControl> Controller { get; }
 
         public static readonly DependencyProperty DisplayColumnsByTitleProperty = DependencyProperty.Register(nameof(DisplayColumnsByTitle), typeof(bool), typeof(QueryBuilderControl), null);
         [Bindable(true)]
@@ -32,6 +34,7 @@ namespace KoS.Apps.SharePoint.SmartCAML.Editor.Controls
         public QueryBuilderControl()
         {
             InitializeComponent();
+            Controller = new OrderedList<QueryFilterControl>(ucFilters);
             AddFilter();
         }
 
@@ -40,53 +43,16 @@ namespace KoS.Apps.SharePoint.SmartCAML.Editor.Controls
             AddFilter();
         }
 
-        private void QueryFilterControl_OnRemoveClick(object sender, EventArgs e)
-        {
-            ucFilters.Children.Remove((UIElement)sender);
-            Changed?.Invoke(this, EventArgs.Empty);
-        }
-
         private void AddFilter()
         {
-            var filterControl = new QueryFilterControl();
-            filterControl.RemoveClick += QueryFilterControl_OnRemoveClick;
-            filterControl.Changed += (sender, args) => Changed?.Invoke(this, EventArgs.Empty);
-            filterControl.Up += MoveFilterUp;
-            filterControl.Down += MoveFilterDown;
-            filterControl.SetBinding(QueryFilterControl.DisplayColumnsByTitleProperty, new Binding
+            Controller
+                .Add()
+                .SetBinding(QueryFilterControl.DisplayColumnsByTitleProperty, new Binding
             {
                 Source = this,
                 Path = new PropertyPath(nameof(DisplayColumnsByTitle)),
                 Mode = BindingMode.TwoWay
             });
-
-            ucFilters.Children.Add(filterControl);
-        }
-
-        private void MoveFilterDown(object control, EventArgs eventArgs)
-        {
-            var filterControl = (UIElement)control;
-            var index = ucFilters.Children.IndexOf(filterControl);
-
-            if (index > 0)
-                MoveFilter(filterControl, index + 1);
-        }
-
-        private void MoveFilterUp(object control, EventArgs eventArgs)
-        {
-            var filterControl = (UIElement)control;
-            var index = ucFilters.Children.IndexOf(filterControl);
-
-            if (index > 0)
-                MoveFilter(filterControl, index - 1);
-        }
-
-        private void MoveFilter(UIElement filterControl, int newIndex)
-        {
-            ucFilters.Children.Remove(filterControl);
-            ucFilters.Children.Insert(newIndex, filterControl);
-
-            Changed?.Invoke(this, EventArgs.Empty);
         }
 
         public ViewBuilder Build()